+++ cmd/command_options.go
@@ -1,7 +1,9 @@
 package cmd
 
 import (
+	"fmt"
 	"github.com/lukso-network/lukso-cli/src/network"
+	"github.com/lukso-network/lukso-cli/src/network/types"
 	"github.com/spf13/cobra"
 )
 
@@ -16,6 +18,11 @@ const (
 	CommandOptionChain     = "chain"
 	CommandOptionNodeName  = "nodeName"
 	CommandOptionDevConfig = "devConfig"
+
+	CommandOptionFrom      = "from"
+	CommandOptionFromShort = "f"
+	CommandOptionTo        = "to"
+	CommandOptionToShort   = "t"
 )
 
 func readConsensusApiEndpoint(cmd *cobra.Command, nodeConf *network.NodeConfigs) (string, error) {
@@ -55,4 +62,33 @@ func readDepositAddress(cmd *cobra.Command, nodeConf *network.NodeConfigs) (stri
 	}
 
 	return nodeConf.DepositDetails.ContractAddress, nil
+}
+
+func readRangeFromCommand(cmd *cobra.Command) (types.ValidatorRange, error) {
+	vRange := types.ValidatorRange{}
+	from, err := cmd.Flags().GetInt64(CommandOptionFrom)
+	if err != nil {
+		return vRange, err
+	}
+	to, err := cmd.Flags().GetInt64(CommandOptionTo)
+	if err != nil {
+		return vRange, err
+	}
+
+	// default to 0
+	if from == -1 {
+		from = 0
+	}
+
+	if to == -1 {
+		return vRange, fmt.Errorf("--to not given")
+	}
+
+	if to <= from {
+		return vRange, fmt.Errorf("--to (%d) must be greater than --from (%d)", to, from)
+	}
+
+	vRange.From = from
+	vRange.To = to
+	return vRange, nil
 }