+++ main.go
@@ -4,7 +4,9 @@ Copyright © 2022 NAME HERE <EMAIL ADDRESS>
 */
 package main
 
-import "github.com/lukso-network/lukso-cli/cmd"
+import (
+	"github.com/lukso-network/lukso-cli/cmd"
+)
 
 func main() {
 	cmd.Execute()