+++ src/network/types/validator_keys.go
@@ -1,0 +1,6 @@
+package types
+
+type ValidatorRange struct {
+	From int64
+	To   int64
+}