+++ src/network/validator_secrets.go
@@ -1,6 +1,7 @@
 package network
 
 import (
+	"github.com/lukso-network/lukso-cli/src/utils"
 	"gopkg.in/yaml.v3"
 	"os"
 )
@@ -15,14 +16,8 @@ type DepositDetails struct {
 type ValidatorCredentials struct {
 	ValidatorMnemonic  string `yaml:""`
 	WithdrawalMnemonic string `yaml:""`
-}
-
-type ValidatorSecretsV0 struct {
-	ValidatorMnemonic  string `yaml:""`
-	WithdrawalMnemonic string `yaml:""`
-	ForkVersion        string `yaml:""`
-
-	Deposit *DepositDetails `yaml:""`
+	ValidatorIndexFrom int64  `yaml:""`
+	ValidatorIndexTo   int64  `yaml:""`
 }
 
 func (valSec *ValidatorCredentials) WriteToFile(fileName string) error {
@@ -31,4 +26,11 @@ func (valSec *ValidatorCredentials) WriteToFile(fileName string) error {
 		return err
 	}
 	return os.WriteFile(fileName, rawData, os.ModePerm)
+}
+
+func (valSec *ValidatorCredentials) Print() {
+	utils.ColoredPrintln("ValidatorMnemonic:", valSec.ValidatorMnemonic)
+	utils.ColoredPrintln("Withdrawal Mnemonic:", valSec.WithdrawalMnemonic)
+	utils.ColoredPrintln("Validator Index From:", valSec.ValidatorIndexFrom)
+	utils.ColoredPrintln("Validator Index To:", valSec.ValidatorIndexTo)
 }